[tempoross] Douse all fires in phases past phase 1
If we have a bad start we could get another cloud event meaning we get 4
more fires in a later phase. This commit adds a check for the case where
we doused all fires on the island but the single fire on the boat
doesn't get picked up because it's out of range. We simply check for
having exactly 2 water buckets in our inventory and if that's the cause
we should move to the boat and douse the fire there.