[birdhouse] Fix pathfinding issue
Pathfinding from the last bird house to the little island in the north
results in a path using the southern mushroom tree. This means we take a
lot of fall damage and it's probably slower than just walking. We add 2
intermediate destinations to avoid this and to avoid pathfinding using a
digsite teleport.