[cooking] Fix double pickup and optimize sonic cook
Due to the dropping method trying to drop everything on 1 tick the
plugin would occassionly make an error and pick up two food pieces when
starting a new inventory. We now cleanly drop the inventory over 3 ticks
(fastest possible, 27 drop actions with a max of 10 actions per tick).
We also keep 1 food in the inventory so we can immediately start cooking
after dropping.